+++ x/auth/keeper/deterministic_test.go
@@ -202,7 +202,7 @@ func (suite *DeterministicTestSuite) TestGRPCQueryParameters() {
 			rapid.Uint64Min(1).Draw(t, "sig-verify-cost-ed25519"),
 			rapid.Uint64Min(1).Draw(t, "sig-verify-cost-Secp256k1"),
 		)
-		err := suite.accountKeeper.SetParams(suite.ctx, params)
+		err := suite.accountKeeper.Params.Set(suite.ctx, params)
 		suite.Require().NoError(err)
 
 		req := &types.QueryParamsRequest{}
@@ -212,7 +212,7 @@ func (suite *DeterministicTestSuite) TestGRPCQueryParameters() {
 	// Regression test
 	params := types.NewParams(15, 167, 100, 1, 21457)
 
-	err := suite.accountKeeper.SetParams(suite.ctx, params)
+	err := suite.accountKeeper.Params.Set(suite.ctx, params)
 	suite.Require().NoError(err)
 
 	req := &types.QueryParamsRequest{}